requirements.txt
django==4.2.15
django-widget-tweaks==1.5.0
Pillow==9.5.0
Jinja2==3.1.2
numpy==1.24.4
pandas==2.0.3
joblib==1.2.0
XlsxWriter==3.1.0
openpyxl==3.1.2
folium==0.14.0
geopy==2.3.0
crispy-bootstrap4==2023.1  
matplotlib==3.7.1
seaborn==0.12.2
chromadb==0.5.1
tiktoken==0.5.0
protobuf==3.20.3
importlib-metadata==8.0.0
gunicorn==23.0.0
wheel==0.44.0
coolsms_python_sdk==2.0.3
django-model-utils==4.5.1
